refactor: consolidate escape_erlang_string into beamtalk-core
Move the canonical implementation from beamtalk-cli to beamtalk-core so
both the CLI and ffi_type_specs (already in core) share a single copy.
The CLI re-exports via `pub(crate) use` so all existing call sites are
unchanged. Tests relocated alongside the implementation in util.rs.
