+++ crates/beamtalk-cli/src/beam_compiler.rs
@@ -86,28 +86,7 @@ fn is_runtime_unavailable_error(err: &miette::Report) -> bool {
     msg.starts_with(RUNTIME_UNAVAILABLE_PREFIX)
 }
 
-/// Escapes a string for use in an Erlang string literal.
-///
-/// This escapes backslashes, quotes, and control characters to prevent
-/// injection attacks and ensure valid Erlang term syntax when constructing
-/// Erlang terms from file paths and other user-supplied strings.
-///
-/// Handles: `\`, `"`, `\n`, `\r`, `\t`, `\0`.
-pub(crate) fn escape_erlang_string(s: &str) -> String {
-    let mut result = String::with_capacity(s.len());
-    for c in s.chars() {
-        match c {
-            '\\' => result.push_str("\\\\"),
-            '"' => result.push_str("\\\""),
-            '\n' => result.push_str("\\n"),
-            '\r' => result.push_str("\\r"),
-            '\t' => result.push_str("\\t"),
-            '\0' => result.push_str("\\0"),
-            _ => result.push(c),
-        }
-    }
-    result
-}
+pub(crate) use beamtalk_core::codegen::core_erlang::escape_erlang_string;
 
 /// Validates that a module name contains only safe identifier characters.
 ///
@@ -1239,55 +1218,6 @@ end
         // Check that output directory was created and compilation succeeded
         assert!(output_dir.exists());
         result.expect("compile_batch should succeed when escript is available");
-    }
-
-    // Tests for escape_erlang_string
-    #[test]
-    fn test_escape_erlang_string_empty() {
-        assert_eq!(escape_erlang_string(""), "");
-    }
-
-    #[test]
-    fn test_escape_erlang_string_no_special_chars() {
-        assert_eq!(escape_erlang_string("hello"), "hello");
-        assert_eq!(escape_erlang_string("foo_bar"), "foo_bar");
-        assert_eq!(escape_erlang_string("path/to/file"), "path/to/file");
-    }
-
-    #[test]
-    fn test_escape_erlang_string_backslashes() {
-        assert_eq!(escape_erlang_string("a\\b"), "a\\\\b");
-        assert_eq!(escape_erlang_string("\\\\"), "\\\\\\\\");
-    }
-
-    #[test]
-    fn test_escape_erlang_string_quotes() {
-        assert_eq!(escape_erlang_string("a\"b"), "a\\\"b");
-        assert_eq!(escape_erlang_string("\"test\""), "\\\"test\\\"");
-    }
-
-    #[test]
-    fn test_escape_erlang_string_newlines() {
-        assert_eq!(escape_erlang_string("line1\nline2"), "line1\\nline2");
-        assert_eq!(escape_erlang_string("\r\n"), "\\r\\n");
-    }
-
-    #[test]
-    fn test_escape_erlang_string_tabs() {
-        assert_eq!(escape_erlang_string("col1\tcol2"), "col1\\tcol2");
-    }
-
-    #[test]
-    fn test_escape_erlang_string_null_byte() {
-        assert_eq!(escape_erlang_string("\0"), "\\0");
-    }
-
-    #[test]
-    fn test_escape_erlang_string_combined() {
-        assert_eq!(
-            escape_erlang_string("path\\to\\\"file\"\n"),
-            "path\\\\to\\\\\\\"file\\\"\\n"
-        );
     }
 
     // Tests for is_valid_module_name
